refactor backend code for readability
